+++ src/test/RefundableDutchAuction.t.sol
@@ -8,14 +8,14 @@ import "forge-std/Test.sol";
 import {RefundableDutchAuction} from "../RefundableDutchAuction.sol";
 
 interface IERC721 {
-    function ownerOf(uint tokenId) external view returns (address);
+    function ownerOf(uint256 tokenId) external view returns (address);
 }
 
 contract RefundableDutchAuctionTest is Test {
     RefundableDutchAuction dutch;
 
-    uint public constant DUTCH_AUCTION_START_ID = 0;
-    uint public constant DUTCH_AUCTION_END_ID = 9999;
+    uint256 public constant DUTCH_AUCTION_START_ID = 0;
+    uint256 public constant DUTCH_AUCTION_END_ID = 9999;
 
     function setUp() public {
         dutch = new RefundableDutchAuction(DUTCH_AUCTION_START_ID, DUTCH_AUCTION_END_ID);
